pyproject.toml
[project]
name = "tp2-pipeline"
version = "0.1.0"
description = "Pipeline d'acquisition et transformation de données Open Data"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "duckdb>=0.9.0",
    "httpx>=0.25.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "tqdm>=4.66.0",
    "pyarrow>=14.0.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

